[Seek Handler] fixed: protect access of 'g_application.m_pPlayer' - because we don't have CApplicationPlayer we must be very careful when accessing m_pPlayer

+++ xbmc/utils/SeekHandler.h
@@ -1,7 +1,7 @@
 #pragma once
 /*
- *      Copyright (C) 2012 Team XBMC
- *      http://www.xbmc.org
+ *      Copyright (C) 2012-2013 Team XBMC
+ *      http://xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -14,18 +14,17 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, write to
- *  the Free Software Foundation, 675 Mass Ave, Cambridge, MA 02139, USA.
- *  http://www.gnu.org/copyleft/gpl.html
+ *  along with XBMC; see the file COPYING.  If not, see
+ *  <http://www.gnu.org/licenses/>.
  *
  */
 
 #include <map>
+#include <utility>
 #include <vector>
 
 #include "guilib/Key.h"
 #include "interfaces/IActionListener.h"
-#include "settings/Settings.h"
 #include "settings/lib/ISettingCallback.h"
 #include "threads/CriticalSection.h"
 #include "utils/Stopwatch.h"
@@ -74,7 +73,7 @@ private:
   std::map<SeekType, int > m_seekDelays;
   bool m_requireSeek;
   bool m_analogSeek;
-  int m_seekSize;
+  double m_seekSize;
   int m_seekStep;
   std::map<SeekType, std::vector<int> > m_forwardSeekSteps;
   std::map<SeekType, std::vector<int> > m_backwardSeekSteps;
